refactoring: `int32_t get_object_size_real() const`->`size_t get_object_size_real() const`

+++ src/xtopcom/xdata/src/xtable_bstate.cpp
@@ -158,8 +158,8 @@ bool xtable_bstate_t::set_receiptid_pair(base::xtable_shortid_t sid, const base:
     return ret == xsuccess;
 }
 
-int32_t xtable_bstate_t::get_object_size_real() const {
-    int32_t total_size = sizeof(*this);
+size_t xtable_bstate_t::get_object_size_real() const {
+    size_t total_size = sizeof(*this);
     int32_t cache_receiptid_size = 0;
     if (m_cache_receiptid != nullptr) {
         cache_receiptid_size = m_cache_receiptid->get_object_size_real();
@@ -167,7 +167,7 @@ int32_t xtable_bstate_t::get_object_size_real() const {
 
     total_size += cache_receiptid_size;
 
-    xdbg("------cache size------ xtable_bstate_t total_size:%d this:%d,cache_receiptid_size:%d", total_size, sizeof(*this), cache_receiptid_size);
+    xdbg("------cache size------ xtable_bstate_t total_size:%zu this:%d,cache_receiptid_size:%d", total_size, sizeof(*this), cache_receiptid_size);
     return total_size;
 }
