+++ src/xtopcom/xvledger/src/xvstate.cpp
@@ -1343,12 +1343,12 @@ namespace top
             return false;
         }
 
-        int32_t xvbstate_t::get_object_size_real() const {
-            int32_t total_size = sizeof(*this);
+        size_t xvbstate_t::get_object_size_real() const {
+            size_t total_size = sizeof(*this);
             auto ex_size = get_ex_alloc_size();
             total_size +=
                 get_size(m_last_block_hash) + get_size(m_last_full_block_hash) + get_size(get_xvid_str()) + get_size(get_address()) + get_size(get_storage_key()) + ex_size;
-            xdbg("------cache size------ xvbstate_t total_size:%d this:%d,m_last_block_hash:%d,m_last_full_block_hash:%d,xvid_str:%d,address:%d,storage_key:%d,ex_size:%d",
+            xdbg("------cache size------ xvbstate_t total_size:%zu this:%d,m_last_block_hash:%d,m_last_full_block_hash:%d,xvid_str:%d,address:%d,storage_key:%d,ex_size:%d",
                  total_size,
                  sizeof(*this),
                  get_size(m_last_block_hash),
